T: move run configuration tests into `rustSlowTests` package
